pyproject.toml
[project]
name = "agaie"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appdirs==1.4.4",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "banks==2.2.0",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.4",
    "build==1.3.0",
    "cachetools==5.5.2",
    "certifi==2025.7.14",
    "charset-normalizer==3.4.2",
    "chromadb==1.0.16",
    "click==8.2.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "dataclasses-json==0.6.7",
    "datasets==4.0.0",
    "defusedxml==0.7.1",
    "deprecated==1.2.18",
    "dill==0.3.8",
    "dirtyjson==1.0.8",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "durationpy==0.10",
    "fastapi==0.116.1",
    "feedparser>=6.0.11",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.7.0",
    "fsspec==2025.3.0",
    "google-auth==2.40.3",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.3",
    "griffe==1.9.0",
    "grpcio==1.74.0",
    "h11==0.16.0",
    "h2==4.2.0",
    "hf-xet==1.1.5",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.34.3",
    "humanfriendly==10.0",
    "hyperframe==6.1.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "kubernetes==33.1.0",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.72",
    "langchain-openai==0.3.28",
    "langchain-text-splitters==0.3.9",
    "langgraph==0.6.2",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.2",
    "langgraph-sdk==0.2.0",
    "langsmith==0.4.9",
    "llama-cloud==0.1.35",
    "llama-cloud-services==0.6.53",
    "llama-index==0.13.0",
    "llama-index-cli==0.5.0",
    "llama-index-core==0.13.0",
    "llama-index-embeddings-huggingface>=0.6.0",
    "llama-index-embeddings-openai==0.5.0",
    "llama-index-indices-managed-llama-cloud==0.9.0",
    "llama-index-instrumentation==0.4.0",
    "llama-index-llms-openai==0.5.0",
    "llama-index-readers-file==0.5.0",
    "llama-index-readers-llama-parse==0.5.0",
    "llama-index-vector-stores-chroma==0.5.0",
    "llama-index-vector-stores-weaviate>=1.4.1",
    "llama-index-workflows==1.2.0",
    "llama-parse==0.6.53",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.6.3",
    "multiprocess==0.70.16",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "nltk==3.9.1",
    "numexpr>=2.13.0",
    "numpy==2.3.2",
    "oauthlib==3.3.1",
    "onnxruntime==1.22.1",
    "openai==1.98.0",
    "opentelemetry-api==1.36.0",
    "opentelemetry-exporter-otlp-proto-common==1.36.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.36.0",
    "opentelemetry-proto==1.36.0",
    "opentelemetry-sdk==1.36.0",
    "opentelemetry-semantic-conventions==0.57b0",
    "orjson==3.11.1",
    "ormsgpack==1.10.0",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.2.3",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "portalocker==3.2.0",
    "posthog==5.4.0",
    "propcache==0.3.2",
    "protobuf==6.31.1",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pypdf==5.9.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "qdrant-client==1.15.1",
    "ragas==0.3.0",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.1.0",
    "rpds-py==0.27.0",
    "rsa==4.9.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.42",
    "starlette==0.47.2",
    "striprtf==0.0.26",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tokenizers==0.21.4",
    "tqdm==4.67.1",
    "trafilatura>=2.0.0",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "weaviate-client>=4.16.10",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.23.0",
]


[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
where = ["src"]
